feat: restructure component
