+++ src/modules/auth/components/logout-button.tsx
